student_projects/Zielinski_Korwek_Wojtaszek/pyproject.toml
[project]
name = "nlp"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.2",
    "ipykernel>=7.1.0",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "wordcloud>=1.9.5",
]
